first-steps/requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
MessageHandler==1.0.2
pillow==12.0.0
python-dotenv==1.2.1
python-telegram-bot==20.7
reportlab==4.4.4
sniffio==1.3.1
style==1.1.0
typing_extensions==4.15.0
update==0.0.1

